fix: wrong config marshaling (int64 instead of int)
also minor refactor the regex named group matching

+++ config/schema.go
@@ -42,9 +42,9 @@ type configDownload struct {
 	Format       *registered[string, libmangal.Format]
 	Path         *registered[string, string]
 	UserAgent    *registered[string, string]
-	ModeDir      *registered[int, fs.FileMode]
-	ModeFile     *registered[int, fs.FileMode]
-	ModeDB       *registered[int, fs.FileMode]
+	ModeDir      *registered[int64, fs.FileMode]
+	ModeFile     *registered[int64, fs.FileMode]
+	ModeDB       *registered[int64, fs.FileMode]
 	Strict       *registered[bool, bool]
 	SkipIfExists *registered[bool, bool]
 	Provider     configDownloadProvider
